pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "jarvis-cli"
version = "0.1.0"
description = "Local-first AI operator console"
authors = [{name = "Your Name"}]
requires-python = ">=3.11"
dependencies = [
    "typer>=0.9.0",
    "rich>=13.0.0",
    "pydantic>=2.0.0",
    "httpx>=0.25.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "streamlit>=1.28.0",
]

[project.scripts]
jarvis = "app.main:cli"

[tool.setuptools.packages.find]
where = ["."]